Add files via upload
The goal of this project is to assist a financial institution in identifying customers with a lower likelihood of defaulting on loans. By building a predictive model, our data science team aims to classify customers into those who are likely to be good borrowers. Based on the model's insights, we will also develop a set of questions to be asked when customers apply for a loan, enhancing the loan vetting process.